requirements.txt
python-dotenv
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
openai
requests